+++ sync-for-payables/CodatSyncPayables/Models/Operations/ListCompaniesResponse.cs
@@ -20,7 +20,7 @@ namespace Codat.Sync.Payables.Models.Operations
         /// <summary>
         /// OK
         /// </summary>
-        public Companies? Companies { get; set; }
+        public Models.Shared.Companies? Companies { get; set; }
 
         /// <summary>
         /// HTTP response content type for this operation