+++ sync-for-payables/RELEASES.md
@@ -87,3 +87,13 @@ Based on:
 - [csharp v3.0.0] sync-for-payables
 ### Releases
 - [NuGet v3.0.0] https://www.nuget.org/packages/Codat.Sync.Payables/3.0.0 - sync-for-payables
+
+## 2023-12-04 01:29:46
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.125.0 (2.209.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v3.0.1] sync-for-payables
+### Releases
+- [NuGet v3.0.1] https://www.nuget.org/packages/Codat.Sync.Payables/3.0.1 - sync-for-payables